pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where   = ["src"]
include = ["sam3d", "sam3d.*", "automask", "automask.*"]

[project]
name = "SAM3D"
version = "0.2.0"
requires-python = ">=3.10,<3.11"
description = "Improved and CUDA 12.1 compatible SAM3D environment"
readme = "README.md"
dependencies = [
    "hydra-core",
    "omegaconf",
    "absl-py==1.4.0",
    "addict==2.4.0",
    "asttokens==2.2.1",
    "astunparse==1.6.3",
    "attrs==22.2.0",
    "backcall==0.2.0",
    "brotlipy==0.7.0",
    "cachetools==5.3.0",
    "certifi",
    "cffi",
    "charset-normalizer",
    "click==8.1.3",
    "clip",
    "comm==0.1.3",
    "ConfigArgParse==1.5.3",
    "contourpy==1.0.7",
    "cryptography",
    "cycler==0.11.0",
    "dash==2.9.2",
    "dash-core-components==2.0.0",
    "dash-html-components==2.0.0",
    "dash-table==5.0.0",
    "debugpy==1.6.7",
    "decorator==5.1.1",
    "executing==1.2.0",
    "fastjsonschema==2.16.3",
    "Flask==2.2.3",
    "flatbuffers==23.3.3",
    "flit_core",
    "fonttools==4.39.3",
    "ftfy==6.1.1",
    "gast==0.4.0",
    "google-auth",
    "google-auth-oauthlib",
    "google-pasta",
    "grpcio==1.53.0",
    "h5py",
    "idna",
    "imageio==2.27.0",
    "importlib-metadata==6.2.0",
    "importlib-resources==5.12.0",
    "ipykernel==6.22.0",
    "ipython==8.12.0",
    "ipywidgets==8.0.6",
    "itsdangerous==2.1.2",
    "jax==0.4.8",
    "jedi==0.18.2",
    "Jinja2==3.1.2",
    "joblib==1.2.0",
    "jsonschema==4.17.3",
    "jupyter_client==8.1.0",
    "jupyter_core==5.3.0",
    "jupyterlab-widgets==3.0.7",
    "keras",
    "kiwisolver==1.4.4",
    "lazy_loader==0.2",
    "libclang==16.0.0",
    "Markdown==3.4.3",
    "MarkupSafe==2.1.2",
    "matplotlib==3.7.1",
    "matplotlib-inline==0.1.6",
    "ml-dtypes",
    "nbformat==5.7.0",
    "nest-asyncio==1.5.6",
    "networkx==3.1",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "open3d==0.17.0",
    "opencv-python==4.7.0.72",
    "opt-einsum==3.3.0",
    "packaging==23.0",
    "pandas==2.0.0",
    "parso==0.8.3",
    "pexpect==4.8.0",
    "pickleshare==0.7.5",
    "Pillow==9.4.0",
    "pkgutil_resolve_name==1.3.10",
    "platformdirs==3.2.0",
    "plotly==5.14.1",
    "prompt-toolkit==3.0.38",
    "protobuf",
    "psutil==5.9.4",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.2",
    "pyasn1==0.4.8",
    "pyasn1-modules==0.2.8",
    "pycparser",
    "Pygments==2.14.0",
    "pyOpenSSL",
    "pyparsing==3.0.9",
    "pyquaternion==0.9.9",
    "pyrsistent==0.19.3",
    "PySocks",
    "python-dateutil==2.8.2",
    "pytz==2023.3",
    "PyWavelets==1.4.1",
    "PyYAML==6.0",
    "pyzmq==25.0.2",
    "regex==2023.3.23",
    "requests",
    "requests-oauthlib==1.3.1",
    "rsa==4.9",
    "scikit-image==0.20.0",
    "scikit-learn==1.2.2",
    "scipy==1.15.3",
    "segment-anything",
    "six",
    "stack-data==0.6.2",
    "tenacity==8.2.2",
    "tensorboard",
    "tensorboard-data-server",
    "tensorboard-plugin-wit",
    "tensorflow",
    "tensorflow-estimator",
    "tensorflow-io-gcs-filesystem",
    "termcolor==2.2.0",
    "threadpoolctl==3.1.0",
    "tifffile==2023.4.12",
    "torch",
    "torch-scatter",
    "torchaudio",
    "torchvision",
    "tornado==6.2",
    "tqdm==4.65.0",
    "traitlets==5.9.0",
    "typing_extensions",
    "tzdata==2023.3",
    "urllib3",
    "wcwidth==0.2.6",
    "Werkzeug==2.2.3",
    "widgetsnbextension==4.0.7",
    "wrapt==1.14.1",
    "zipp==3.15.0",
    "plyfile>=1.1.2",
]

[project.scripts]
sam3d = "sam3d.sam3d:main"
format = "sam3d.data_util.format_for_is:main"
gt_format = "sam3d.data_util.format_gt:main"
prepare_data = "sam3d.data_util.prepare_data_ist:main"
eval_instance = "sam3d.evaluation.evaluate_instance_seg:main"